% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_evento.R
\name{log_evento}
\alias{log_evento}
\title{Log Evento - Mensagem Única}
\usage{
log_evento(caminho, msg = "", script = "")
}
\arguments{
\item{caminho}{Caminho onde o log deve ser salvo - string}

\item{msg}{Mensagem que deve ser salva no arquivo log - string
Default: ""}

\item{script}{Nome do script onde a função foi invocada. - string
Esse parâmetro serve para compor o nome do arquivo log que será gerado
Default: ""}
}
\value{
A função salva um arquivo log com cabeçalho e com a mensagem escolhida
O nome do arqivo gerado é composto da data e hora em que a função foi invocada e do parâmetro script
Nome do arquivo log gerado: aammdd-hhmmss-script.log
}
\description{
Log Evento - Mensagem Única
}
