% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_eventos.R
\name{log_eventos}
\alias{log_eventos}
\title{Log Eventos - Múltipas Mensagens}
\usage{
log_eventos(caminho, msgs = c(""), script = "")
}
\arguments{
\item{caminho}{Caminho onde o log deve ser salvo - string}

\item{msgs}{Mensagens que devem ser salvas no arquivo log - list
Default: ""}

\item{script}{Nome do script onde a função foi invocada. - string
Esse parâmetro serve para compor o nome do arquivo log que será gerado
Default: ""}
}
\value{
A função salva um arquivo log com cabeçalho e com as mensagens escolhidas
O nome do arqivo gerado é composto da data e hora em que a função foi invocada e do parâmetro script
Nome do arquivo log gerado: aammdd-hhmmss-script.log
}
\description{
Log Eventos - Múltipas Mensagens
}
